add 'Copy signal' btn for all activities
