+++ src/components/SignalCard/controls/CopySignal.js
@@ -4,8 +4,14 @@ import debounce from 'lodash.debounce'
 import Button from '@santiment-network/ui/Button'
 import { createTrigger } from '../../../ducks/Signals/common/actions'
 import styles from './CopySignal.module.scss'
+import { checkIsLoggedIn } from '../../../pages/UserSelectors'
+import isEqual from 'lodash.isequal'
 
-const CopySignal = ({ signal, createTrigger }) => {
+const CopySignal = ({ isAuthor, isCreated, signal, createTrigger }) => {
+  if (isAuthor || isCreated) {
+    return null
+  }
+
   const copySignal = debounce(() => {
     const newSignal = { ...signal }
     delete newSignal.id
@@ -33,7 +39,29 @@ const mapDispatchToProps = dispatch => ({
   }
 })
 
+const mapStateToProps = (state, { creatorId, signal }) => {
+  const isLoggedIn = checkIsLoggedIn(state)
+
+  return {
+    isAuthor:
+      state &&
+      state.user &&
+      state.user.data &&
+      +state.user.data.id === +creatorId,
+    isLoggedIn: isLoggedIn,
+    isCreated:
+      !isLoggedIn ||
+      (state &&
+        state.signals.all &&
+        state.signals.all.some(
+          item =>
+            item.title === signal.title &&
+            isEqual(signal.settings.operation, item.settings.operation)
+        ))
+  }
+}
+
 export default connect(
-  null,
+  mapStateToProps,
   mapDispatchToProps
 )(CopySignal)
