+++ src/pages/SonarFeed/ActivityRenderer/ActivityWithBacktesting.js
@@ -7,6 +7,7 @@ import { SIGNAL_ANCHORS } from '../../../ducks/Signals/common/constants'
 import { dateDifferenceInWordsString } from '../../../utils/dates'
 import SignalPreview from '../../../ducks/Signals/chart/SignalPreview'
 import styles from './ActivityRenderer.module.scss'
+import CopySignal from '../../../components/SignalCard/controls/CopySignal'
 
 const ActivityWithBacktesting = ({
   index,
@@ -19,51 +20,49 @@ const ActivityWithBacktesting = ({
     settings: { type }
   } = trigger
   return (
-    <div
-      className={cx(
-        styles.activityItem,
-        styles.activityItemBacktest,
-        classes.activityItem,
-        index === 0 && classes.firstActivity
-      )}
-    >
-      <div
-        className={cx(
-          styles.description,
-          styles.activityCustom,
-          styles.activityBacktest
-        )}
-      >
-        <SignalTypeIcon type={type} />
+    <>
+      <SignalTypeIcon type={type} className={styles.icon} />
+      <div>
+        <div className={styles.info}>
+          <div
+            className={cx(
+              styles.description,
+              styles.activityCustom,
+              styles.activityBacktest
+            )}
+          >
+            <div className={styles.center}>
+              <h4 className={styles.title}>
+                <Link
+                  to={`/sonar/signal/${signalId}${SIGNAL_ANCHORS.ACTIVITIES}`}
+                  className={styles.link}
+                >
+                  {title}
+                </Link>
+              </h4>
+              <Markdown
+                source={Object.values(payload)[0]}
+                className={classes.activityMarkdown}
+              />
 
-        <div className={styles.center}>
-          <h4 className={styles.title}>
-            <Link
-              to={`/sonar/signal/${signalId}${SIGNAL_ANCHORS.ACTIVITIES}`}
-              className={styles.link}
-            >
-              {title}
-            </Link>
-          </h4>
-          <Markdown
-            source={Object.values(payload)[0]}
-            className={classes.activityMarkdown}
-          />
-
-          <span className={styles.dateBacktest}>
-            {dateDifferenceInWordsString(triggeredAt)}
-          </span>
+              <span className={styles.dateBacktest}>
+                {dateDifferenceInWordsString(triggeredAt)}
+              </span>
+            </div>
+          </div>
+          <div className={styles.preview}>
+            <SignalPreview
+              trigger={trigger}
+              type={type}
+              showExpand={false}
+              showTitle={false}
+            />
+          </div>
         </div>
+
+        <CopySignal signal={trigger} />
       </div>
-      <div className={styles.preview}>
-        <SignalPreview
-          trigger={trigger}
-          type={type}
-          showExpand={false}
-          showTitle={false}
-        />
-      </div>
-    </div>
+    </>
   )
 }
