+++ src/pages/feed/GeneralFeed/FeedItemRenderer/FeedItemRenderer.module.scss
@@ -18,6 +18,7 @@
 
 .activityItem {
   margin-bottom: 16px;
+  display: flex;
 }
 
 .firstActivity {
